requirements.txt
python==3.9.6
django==3.1.8=pypi_0
django-cors-headers==3.7.0
djangorestframework==3.12.0
pyjwt==2.1.0
openai==0.11.0